Snapshot build D0000169

+++ BB3/App/gui/tasks/menu/system/datetime/datetime.c
@@ -29,7 +29,7 @@ REGISTER_TASK_IL(datetime,
     lv_obj_t * gnss;
 );
 
-static bool datetime_cb(lv_obj_t * obj, lv_event_t event, uint8_t index)
+static bool datetime_cb(lv_obj_t * obj, lv_event_t event, uint16_t index)
 {
 	if (event == LV_EVENT_VALUE_CHANGED)
 	{
